-robot vs tablet data ready for statistaical analysis -Granger- needed sanity check

+++ gui.py
@@ -22,8 +22,8 @@ class Label(QLabel):
         point = QPoint(0,0)
         scaledPix = self.pixmap.scaled(size, Qt.KeepAspectRatio, transformMode = Qt.SmoothTransformation)
         # start painting the label from left upper corner
-        point.setX((size.width() - scaledPix.width())/2)
-        point.setY((size.height() - scaledPix.height())/2)
+        point.setX(int((size.width() - scaledPix.width())/2))
+        point.setY(int((size.height() - scaledPix.height())/2))
         painter.drawPixmap(point, scaledPix)
 
 class CheckableComboBox(QComboBox):
